Add default values for kubeconfig
